remove sha256_ethereum.hpp
templated functions can only be in header files
